Move submittedProposals selector logic to app
