+++ src/selectors/api.js
@@ -1,9 +1,8 @@
 import get from "lodash/fp/get";
 import eq from "lodash/fp/eq";
 import filter from "lodash/fp/filter";
-import find from "lodash/fp/find";
 import compose from "lodash/fp/compose";
-import { and, or, bool, constant, not } from "../lib/fp";
+import { and, or, bool, constant } from "../lib/fp";
 import { PROPOSAL_STATUS_UNREVIEWED, PROPOSAL_STATUS_CENSORED } from "../constants";
 
 const getIsApiRequesting = key => bool(get(["api", key, "isRequesting"]));
@@ -110,10 +109,7 @@ export const unreviewedProposals = filtered(PROPOSAL_STATUS_UNREVIEWED);
 export const censoredProposals = filtered(PROPOSAL_STATUS_CENSORED);
 export const unvettedProposalsIsRequesting = or(isApiRequestingInit, isApiRequestingUnvetted);
 export const unvettedProposalsError = or(apiInitError, apiUnvettedError);
-export const proposal = or(compose(get("proposal"), apiProposalResponse), (state) => {
-  const submittedProposals = state.app.submittedProposals;
-  return submittedProposals[submittedProposals.lastSubmitted];
-}, constant({}));
+export const proposal = compose(get("proposal"), apiProposalResponse);
 export const proposalToken = compose(get(["censorshiprecord", "token"]), proposal);
 export const proposalComments = or(compose(get("comments"), apiProposalCommentsResponse), constant([]));
 export const proposalIsRequesting = or(isApiRequestingInit, isApiRequestingProposal);
@@ -128,8 +124,4 @@ export const newProposalName = compose(get("name"), apiNewProposalPayload);
 export const newProposalDescription = compose(get("description"), apiNewProposalPayload);
 export const newProposalFiles = compose(get("files"), apiNewProposalPayload);
 export const setStatusProposal = compose(get("status"), apiSetStatusProposalResponse);
-export const getProposalFiles = compose(get("files"), proposal);
-export const isMarkdown = compose(eq("index.md"), get("name"));
-export const getMarkdownFile = compose(find((isMarkdown)), getProposalFiles);
-export const getNotMarkdownFile = compose(filter(not(isMarkdown)), getProposalFiles);
 export const redirectedFrom = get(["api", "login", "redirectedFrom"]);
